test(event cache): rewrite the `test_redact_touches_thread` to make it resilient to races
In the previous version of the thread, the following sequence of events
could happen:

- we subscribe to the thread linked chunk changes
- *then*, the events are being added to the thread linked chunk

This is an edge case where, since we've subscribed to the thread linked
chunk, the thread root will be "known" to be part of a thread, and will
be appended to the thread linked chunk.

If the events happen in the other order (first the events are added to
the thread linked chunk, then we subscribe to the changes), then the
thread root will not be part of the thread linked chunk (because when it
arrived, we didn't know it would be a thread root). As such, the thread
linked chunk state would end up being different in this case.

The solution is to make it so that the thread linked chunk is always
subscribed to *before* any events are added to it. This way, we make
sure that we'll always have the thread root in the thread linked chunk.