fix: `NotificationSettings::unmute_room` didn't clear the cached notification mode
As a result, the clients still received the cached value in the `RoomInfo`s.

A way to solve this could be to clear this value in `unmute_room` or some other point of this flow, but `SlidingSyncResponseProcessor::update_in_memory_caches` already updates this value *if* the received room has a custom notification mode, it just never clears it, so maybe it's the best place to also add its counterpart.